% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfAppleCOVID19MobilityData}
\alias{dfAppleCOVID19MobilityData}
\title{Apple COVID-19 Mobility data}
\format{
A data frame
\describe{
    \item{GeoType}{Geo type}
    \item{Region}{Region name}
    \item{TransportationType}{One of "driving", "transit", "walking".}
    \item{AlternativeName}{Alternative name}
    \item{SubRegion}{Sub region name}
    \item{Country}{Country}
    \item{AbsoluteTime}{Absolute time of the date in seconds from 1900-01-01.}
    \item{Value}{The value of the stock.}
    \item{Date}{The POSIXct date corresponding to \code{dfAppleCOVID19MobilityData$AbsoluteTime}.}
}
}
\source{
\url{https://covid19.apple.com/mobility}
}
\usage{
dfAppleCOVID19MobilityData
}
\description{
Apple COVID-19 Mobility data sample with data for Nice, France only.
}
\details{
The data from 2020-01-13 to 2021-01-15.
}
\seealso{
Other MathematicaVsR data: 
\code{\link{dfDistributionData}},
\code{\link{dfFinancialData}},
\code{\link{dfMushroom}},
\code{\link{dfSP500}},
\code{\link{dfTemperatureData}},
\code{\link{dfTitanic}},
\code{\link{dfUSAPresidentsSpeeches}},
\code{\link{textHamlet}}
}
\concept{MathematicaVsR data}
\keyword{datasets}
