% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfFinancialData}
\alias{dfFinancialData}
\title{GE stock data}
\format{A data frame with 1258 rows and 2 columns.
\describe{
    \item{Time}{Date.}
    \item{Value}{Adjusted stock price in US dollars.}
}}
\usage{
dfFinancialData
}
\description{
GE stock data from 2014-01-01 to 2019-01-01.
}
\seealso{
Other MathematicaVsR data: \code{\link{dfDistributionData}},
  \code{\link{dfMushroom}}, \code{\link{dfSP500}},
  \code{\link{dfTemperatureData}}, \code{\link{dfTitanic}},
  \code{\link{dfUSAPresidentsSpeeches}},
  \code{\link{textHamlet}}
}
\concept{MathematicaVsR data}
\keyword{datasets}
