% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfSP500}
\alias{dfSP500}
\title{S&P 500 stock daily closing values}
\format{
A data frame
\describe{
    \item{StockID}{Stock identifier.}
    \item{AbsoluteTime}{Absolute time of the date in seconds from 1900-01-01.}
    \item{Value}{The value of the stock.}
    \item{Date}{The POSIXct date corresponding to \code{dfSP500$AbsoluteTime}.}
}
}
\source{
\url{https://reference.wolfram.com/language/ref/FinancialData.html}
}
\usage{
dfSP500
}
\description{
S&P 500 stock daily closing values retrieved with Mathematica's function \code{FinancialData}.
}
\details{
For additional details see \url{https://en.wikipedia.org/wiki/S&P_500_Index} and
\url{https://en.wikipedia.org/wiki/List_of_S&P_500_companies}.
}
\seealso{
Other MathematicaVsR data: 
\code{\link{dfAppleCOVID19MobilityData}},
\code{\link{dfDistributionData}},
\code{\link{dfFinancialData}},
\code{\link{dfMushroom}},
\code{\link{dfTemperatureData}},
\code{\link{dfTitanic}},
\code{\link{dfUSAPresidentsSpeeches}},
\code{\link{textHamlet}}
}
\concept{MathematicaVsR data}
\keyword{datasets}
