% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfUSAPresidentsSpeeches}
\alias{dfUSAPresidentsSpeeches}
\title{USA presidents speeches}
\format{A data frame}
\source{
\url{http://www.gutenberg.org/ebooks/925}
}
\usage{
dfUSAPresidentsSpeeches
}
\description{
USA presidents speeches together with related metadata.
}
\keyword{datasets}
